requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
coverage==7.6.3
Django==4.2.16
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-select2==8.2.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pydotplus==2.0.2
pyparsing==3.1.4
pyproject==1.3.1
pytz==2024.2
PyYAML==6.0.2
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
